DesktopSessionAgent now hosts the video capturer and provides necessary plumbing to drive it via an IPC channel.
BUG=134694


Review URL: https://chromiumcodereview.appspot.com/11413022

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@169587 0039d316-1c4b-4281-b951-d872f2087c98

+++ remoting/base/shared_buffer.cc
@@ -16,7 +16,7 @@ SharedBuffer::SharedBuffer(uint32 size)
 }
 
 SharedBuffer::SharedBuffer(
-    intptr_t id, base::SharedMemoryHandle handle, uint32 size)
+    int id, base::SharedMemoryHandle handle, uint32 size)
     : id_(id),
       shared_memory_(handle, kReadOnly),
       size_(size) {
@@ -24,7 +24,7 @@ SharedBuffer::SharedBuffer(
 }
 
 SharedBuffer::SharedBuffer(
-    intptr_t id, base::SharedMemoryHandle handle, base::ProcessHandle process,
+    int id, base::SharedMemoryHandle handle, base::ProcessHandle process,
     uint32 size)
     : id_(id),
       shared_memory_(handle, kReadOnly, process),
