+++ remoting/base/shared_buffer_unittest.cc
@@ -9,8 +9,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 const uint32 kBufferSize = 4096;
 const int kPattern = 0x12345678;
 
-const intptr_t kIdZero = 0;
-const intptr_t kIdOne = 1;
+const int kIdZero = 0;
+const int kIdOne = 1;
 
 namespace remoting {
